consolidated outputers into builtin plugin and created new output base class to standardize output pattern.
