+++ supergsl/plugins/builtin/__init__.py
@@ -3,6 +3,8 @@ from supergsl.core.function import SuperGSLFunctionDeclaration
 
 from .fuse import FusionAssembler
 from .print import SuperGSLTypePrintFunction
+from .output.json_output import JSONOutput
+from .output.sbol_output import SBOLOutput
 
 class BuiltinAssemblersPlugin(SuperGSLPlugin):
     """Plugin stub to help register basic Assemblers."""
@@ -14,3 +16,13 @@ class BuiltinAssemblersPlugin(SuperGSLPlugin):
 
         self.register_function('builtin', 'print', SuperGSLFunctionDeclaration(
             SuperGSLTypePrintFunction, compiler_settings))
+
+        self.register_function(
+            'builtin',
+            'output_json',
+            SuperGSLFunctionDeclaration(JSONOutput, compiler_settings))
+
+        self.register_function(
+            'builtin',
+            'output_sbol',
+            SuperGSLFunctionDeclaration(SBOLOutput, compiler_settings))
