+++ supergsl/plugins/builtin/examples/output_json_ex1.json
@@ -1,0 +1,6 @@
+from builtin import fuse, output_json
+from S288C import GAL1, GAL3, GAL7
+
+let assembly = fuse { uGAL1 ; pGAL7 ; gGAL3 ; dGAL1 }
+
+output_json(assembly, 'output.json')
