--- supergsl/plugins/sbol/__init__.py
@@ -1,1 +1,0 @@
-from .sbol_output import SBOLOutput
